refactor: component structure modify
